% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collar.history.R
\name{collar.history}
\alias{collar.history}
\title{collar.history}
\usage{
collar.history(data, idcol, study, capdate, mortdate)
}
\arguments{
\item{data}{name of data.frame for capture data}

\item{idcol}{name of column where animal id is stored}

\item{study}{name of study area to choose for collar history table ("idaho", or "washington")}

\item{capdate}{name of column where capture date is stored}

\item{mortdate}{name of column where mortality date is stored}
}
\value{
Returns a date.frame of history of collars for each animal
}
\description{
Transform capture data to collar history
}
\examples{

}
\keyword{capture}
\keyword{collar}
\keyword{data,}
\keyword{history,}
