% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.bbox.R
\name{create.bbox}
\alias{create.bbox}
\title{create.bbox}
\usage{
create.bbox(inshape, id, buffer = TRUE, buffsize = 1, native = TRUE)
}
\arguments{
\item{inshape}{input shapefile to create bounding box around}

\item{id}{character of id for bounding box}

\item{buffer}{Logical. If TRUE (default), create new shapefile with a specified buffer around it}

\item{buffsize}{numeric vector of number of kilometers to buffer around bounding box}

\item{native}{Logical. If TRUE (default), return shapefile in the native projection, if FALSE, return shapefile in the Albers Equal Area (used to create buffer)}
}
\value{
Returns a SpatialPolygonsDataFrame with bounding box with buffer if buffer = TRUE
}
\description{
create a SpatialPolygonsDataFrame from the bounding box coordinates of a shapefile
}
\keyword{bounding}
\keyword{box,}
\keyword{shapefile}
\keyword{spdf,}
