% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine.random.R
\name{determine.random}
\alias{determine.random}
\title{determine.random}
\usage{
determine.random(
  rasterstack,
  poly,
  pts,
  increase.type,
  increment,
  number.iterations,
  seqmin,
  seqmax,
  inc.ratio,
  out.dir
)
}
\arguments{
\item{rasterstack}{name of raster or raster stack object to evaluate}

\item{poly}{spatialpolygonsdataframe to sample random locations from}

\item{pts}{spatialpointsdataframe of used locations}

\item{increase.type}{whether to increase proportionally (i.e., percentage) or numerically (i.e., integers)}

\item{increment}{if increase.type prop a proportion in decimal format (e.g., for 10 percent 0.10) if increase type number indicate number of points to increase at each iteration}

\item{number.iterations}{number of incremental steps to run through}

\item{seqmin}{if increase.type number, indicate the minimum number of random locations to test}

\item{seqmax}{if increase.type number, indicate the maximum number of random locations to test}

\item{inc.ratio}{Logical. TRUE If you desire the number of random points to be greater than the total number of used points, otherwise FALSE}

\item{out.dir}{path to folder where plot should be saved}
}
\value{
Returns a data.frame with the summarized values, and saves a pdf of plots to out.dir named RandomPoints
}
\description{
determine number of random locations needed to represent true mean of landscape variables
}
\keyword{available,}
\keyword{random,}
\keyword{raster}
\keyword{sample,}
