% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnld.snodas.R
\name{dnld.snodas}
\alias{dnld.snodas}
\title{dnld.snodas}
\usage{
dnld.snodas(root.dir, cropshape, startdate, enddate)
}
\arguments{
\item{root.dir}{path to root directory where subdirectories of .gz files, and SWE, and depth files will be stored}

\item{cropshape}{path to shapefile to crop SNODAS data to}

\item{startdate}{character in "YYYY-MM-DD" format of the beginning date to download SNODAS data}

\item{enddate}{character in "YYYY-MM-DD" format of the beginning date to download SNODAS data}
}
\value{
Returns rasters for each day of SWE and Snow Depth saved in sub directories of root.dir
}
\description{
download raw SNODAS data from server, and transform to SWE and Snow Depth rasters
}
\keyword{SNODAS,}
\keyword{SWE,}
\keyword{download}
\keyword{snowpack,}
