% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.sheep.gps.R
\name{id.sheep.gps}
\alias{id.sheep.gps}
\title{id.sheep.gps}
\usage{
id.sheep.gps(gpspath, sheepdb, tzone, capcol, dateformat, mortcol, extracols)
}
\arguments{
\item{gpspath}{path to gps data}

\item{sheepdb}{path to sheep capture database}

\item{tzone}{desired time zone of gps data: "MST" or "US/Pacific"}

\item{capcol}{name of column where capture/start date exists}

\item{dateformat}{character string of the format that date columns are in}

\item{mortcol}{name of column where mortality date or end date exists}

\item{extracols}{vector of extra columns that should be appended to GPS data. Names in vector MUST match names in lookup table}
}
\value{
Returns a data.frame with all gps data and any extra columns desired
}
\description{
Merge Raw GPS Data with AnimalID, Sex, and Eartag ID
}
\examples{

}
\keyword{ID,}
\keyword{animal}
\keyword{append}
\keyword{capture,}
\keyword{gps,}
