% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IFBFat.R
\name{IFBFat}
\alias{IFBFat}
\title{IFBFat}
\usage{
IFBFat(dbpath, rumpfat, units, bcs)
}
\arguments{
\item{dbpath}{path to location of capture database or where fat measurements are stored}

\item{rumpfat}{name of column where rump fat measurement exists}

\item{units}{character. whether rump fat measurement is in "cm" or "mm"}

\item{bcs}{name of column where body condition score exists}
}
\value{
Returns original capture database with IFBFat column appended
}
\description{
calculate IFBFat for bighorn sheep
}
\examples{
\donttest{fat<-IFBFat(dbpath = 'C:/Desktop/yourdb', rumpfat = 'MaxFat', units = "mm", bcs = 'BCS')}
}
\keyword{IFBFat}
\keyword{condition,}
\keyword{nutritional}
