% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.disease.R
\name{merge.disease}
\alias{merge.disease}
\title{merge.disease}
\usage{
\method{merge}{disease}(disease, sheets, cap, aidcol, merge, mergecol)
}
\arguments{
\item{disease}{path to disease excel workbook}

\item{sheets}{sheets within disease data.frame to be merged with capture data}

\item{cap}{path to capture database}

\item{aidcol}{name of column where animal ID exists in capture data.frame}

\item{merge}{Logical. If TRUE (default), function will merge desired component of lab results to capture data}

\item{mergecol}{name of sheet for desired merge}
}
\value{
Returns a list (if merge = TRUE) with list of sheets with lab results and merged data.frame of selected lab results to capture data. If merge = FALSE, a list of sheets for all lab results will be returned
}
\description{
merge lab results with capture data. This will likely only be used for Idaho and Asotin data (Wyoming is already merged)
}
\keyword{disease,}
\keyword{elisa}
\keyword{lab}
\keyword{movi,}
\keyword{pcr,}
\keyword{results,}
\keyword{test,}
