% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLPartPred.R
\name{MLPartPred}
\alias{MLPartPred}
\title{Parturition Prediction with Machine Learning}
\usage{
MLPartPred(jk, spp)
}
\arguments{
\item{jk}{output of BGBFun}

\item{spp}{species for model. Choices are "Elk", "FMD" (Female mule deer), or "BHS" (bighorn sheep)}
}
\value{
data.frame of probability predictions for 0, 1, 2 (pre birth, neonate <48hrs old, neonate>48hrs old)
}
\description{
Use Machine Learning to predict parturition
}
\examples{
\donttest{hg<-MLPartPred(mdat3, spp = "BHS")}

}
\keyword{prediction}
