% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neonate.survival.R
\name{neonate.survival}
\alias{neonate.survival}
\title{neonate.survival}
\usage{
neonate.survival(dbpath, startdate, enddate, units, cuts, cutoff, fatecol,
  censors, model)
}
\arguments{
\item{dbpath}{path to neonate database}

\item{startdate}{name of column where birth date or capture date is stored}

\item{enddate}{name of column where mortality date is stored}

\item{units}{Units desired for modeling survival (e.g. days, weeks, etc)}

\item{cuts}{Logical. TRUE/FALSE. Whether you desire to use a cutoff date for survival}

\item{cutoff}{days from birth to use as cutoff (e.g. 140 days, indicate 140)}

\item{fatecol}{name of column where fate/cause of death is stored}

\item{censors}{Vector of causes in your data.frame in which events should be censored (i.e. CollarFailure)}

\item{model}{Logical. TRUE/FALSE. Whether model of survival (stratified by year) should be done}
}
\value{
Returns a list that includes original dataframe with time alive, and indicator of event as well as the model of survival (if model == TRUE)
}
\description{
Convert capture data of neonates into the proper format for conducting survival analysis. This function will return 2 columns of data: survival time of each in
whatever units desired (days, weeks, months, etc.) and an indicator column of whether the event (i.e., mortality) occurred in binary format. Further, this function allows the option to
use a cutoff age for modeling survival. For instance, in many cases survival of neonatal ungulates is modeled to 140 days.
}
\examples{
\donttest{neo.survival<-neonate.survival(dbpath = yourpath, units = "days", cuts = TRUE, cutoff = 140, censors = "collar_failure", model = TRUE)}

}
\keyword{kaplan}
\keyword{meier}
\keyword{neonate,}
\keyword{survival,}
