% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrettyData.R
\name{PrettyData}
\alias{PrettyData}
\title{PrettyData}
\usage{
PrettyData(dat, idl, filen)
}
\arguments{
\item{ti}{data.frame output from adehabitatLT as.ltraj function. Also the output from trajfun}

\item{data}{data.frame used to make ti}
}
\description{
Create tables of most recent locations to include in report
}
\keyword{bind}
