% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunPartModel.R
\name{RunPartModel}
\alias{RunPartModel}
\title{RunPartModel}
\usage{
RunPartModel(
  gpsdat,
  lookup,
  gpsproj,
  projectedproj,
  subsetmonth,
  tempdir,
  ncpus,
  markdownpath,
  from = NA,
  to = NA,
  subject = NA,
  SP = NA,
  progpath = NA,
  username = NA,
  password = NA,
  email = "no",
  ST = TRUE,
  STUser = NULL,
  STPass = NULL,
  systype = "Win"
)
}
\arguments{
\item{gpsdat}{data.frame of gps data to run model that includes the following columns (in this order): 'IdCol', 'SerialNumber', 'TelemDate', 'x', 'y'. Optional columns include: 'dop', 'X2D.3D', but recommended if available to help improve data cleaning process. IdCol = column that represents unique identifier, TelemDate = as.POSIXct format. Fix time and date, x = x coordinate, y = y coordinate. See data(sheep.gps) for example of expected format.}

\item{lookup}{data.frame of lookup table. This table should include the following columns: 'Frequency', 'Serial', 'IdCol'. Frequency = collar frequency in XXX.XXX format, Serial = Serial number of collar, IdCol = Unique identifier column. MUST match the IdCol in the gpsdat, BirthDate (optional), used to fill in birth dates as they occur and add to plots. Other columns such as VitFrequency, TagNumber are allowed. See data(lookup) for example of expected format.}

\item{gpsproj}{proj4 of gps data}

\item{projectedproj}{proj4 of projected coordinate system to use for data cleaning (UTM or Albers Equal Area recommended)}

\item{subsetmonth}{character object indicating which month to use as the start date for analysis. For example, for a start date of March 1 of current year, use "03"}

\item{tempdir}{path of directory for products to be saved}

\item{ncpus}{number of CPUs for parallel processing. Recommend 1-2 less than max.}

\item{markdownpath}{complete path to location of markdown file for parturition report}

\item{from}{character vector of email sender}

\item{to}{character vector of email recipients}
}
\value{
Resulting object is a pdf with movement metrics for each unique animal, and a data.frame with machine learning predictions
}
\description{
Calculate movement statistics including movement rate, first passage time, bivariate gaussian bridge, and machine learning predictions to identify parturition events of ungulates.
}
\examples{
\donttest{RunPartModel(data=gps, lookup = lookup, gpsproj = +proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs", projectedproj = +proj=utm +zone=11 +ellps=GRS80 +datum=NAD83 +units=m +no_defs", subsetmonth = "01", tempdir = "C:/Users/khuggler/Desktop/", ncpus = 5, markdownpath = "C:/Users/Desktop/PartPlots2.Rmd")}
}
\keyword{bridge,}
\keyword{first}
\keyword{gaussian}
\keyword{learning}
\keyword{machine}
\keyword{movement}
\keyword{parturition,}
\keyword{passage}
\keyword{rate,}
\keyword{time,}
