% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.transects.R
\name{sample.transect}
\alias{sample.transect}
\title{sample.transects}
\usage{
sample.transect(ngroups, nsamps, maxdist)
}
\arguments{
\item{ngroups}{number of sheep groups to sample within}

\item{nsamps}{number of sample distance/azimuth combinations to create for each group}

\item{maxdist}{maximum distance (in meters) to sample from central group location}
}
\value{
Returns data.frame with sampled distance and azimuths for specified number of groups
}
\description{
sample random distance and azimuths for vegetation transects
}
\keyword{sample,}
\keyword{transects,}
\keyword{veg}
