% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheep.activity.R
\name{sheep.activity}
\alias{sheep.activity}
\title{sheep.activity}
\usage{
sheep.activity(activity.fold, collarhistory)
}
\arguments{
\item{activity.fold}{path to activity data}

\item{collarhistory}{path to collar history file}
}
\value{
Returns a data.frame with all activity data and appended animal ID
}
\description{
Merge Animal IDs to activity data
}
\examples{

}
\keyword{ID,}
\keyword{activity,}
\keyword{animal}
\keyword{append}
\keyword{capture,}
