% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheep.gps.R
\name{sheep.gps}
\alias{sheep.gps}
\title{sheep.gps}
\usage{
sheep.gps(vecpath, sheepdb, tzone, serialcol, capcol, dateformat, mortcol,
  extracols)
}
\arguments{
\item{vecpath}{path where vec data is located}

\item{sheepdb}{path to sheep capture database}

\item{tzone}{desired time zone of gps data: "MST" or "US/Pacific"}

\item{serialcol}{name of column in capture datebase/lookup table where Serial Number exists}

\item{capcol}{name of column where capture/start date exists}

\item{dateformat}{character string of the format that date columns are in}

\item{mortcol}{name of column where mortality date or end date exists}

\item{extracols}{vector of extracolumns that should be appended to GPS data. Names in vector MUCH match named in lookup table}
}
\value{
Returns a data.frame with all gps data and any extra columns desired
}
\description{
Merge Raw GPS Data with AnimalID, Sex, and Eartag ID
}
\examples{

}
\keyword{ID,}
\keyword{animal}
\keyword{append}
\keyword{capture,}
\keyword{gps,}
