% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheep.report.R
\name{sheep.report}
\alias{sheep.report}
\title{sheep.report}
\usage{
sheep.report(keys, sheepdb, tzone, serialcol, capcol, dateformat, mortcol,
  extracols, keep.aid, out.dir)
}
\arguments{
\item{keys}{full path to one or more key files}

\item{sheepdb}{path to sheep capture database}

\item{tzone}{desired time zone of gps data: "MST" or "US/Pacific"}

\item{serialcol}{name of column in capture datebase/lookup table where Serial Number exists}

\item{capcol}{name of column where capture/start date exists}

\item{dateformat}{character string of the format that date columns are in}

\item{mortcol}{name of column where mortality date or end date exists}

\item{extracols}{vector of extra columns that should be appended to GPS data. Names in vector MUST match names in lookup table}

\item{keep.aid}{vector of animal IDs to include in report}

\item{out.dir}{path to folder where kml and html map should be saved}
}
\value{
Returns a kml and html map of most recent location for all desired animals to out.dir
}
\description{
generate report with most recent location of desired animal IDs
}
\keyword{capture,}
\keyword{html,}
\keyword{kml,}
\keyword{map}
