% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terra.climate.R
\name{terra.climate}
\alias{terra.climate}
\title{terra.climate}
\usage{
terra.climate(root.dir, cropshape, months)
}
\arguments{
\item{root.dir}{path of root directory where spatial data is located}

\item{cropshape}{path to shapefile to crop rasters to}

\item{months}{vector of months to subset rasters to. Must be numeric (e.g. c(4,5,6,7,8))}
}
\value{
Returns a list of raster bricks with all climate data
}
\description{
Read in netCDF files and manipulate to fit study area
}
\examples{

}
\keyword{climate,}
\keyword{pdsi,}
\keyword{swe}
\keyword{temperature,}
