% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vitMap.R
\name{vitMap}
\alias{vitMap}
\title{Create Plots for Parturition Markdown}
\usage{
vitMap(locdat, vhist, labels, fold, plotdataPath, hg = NULL)
}
\arguments{
\item{locdat}{movement data}

\item{vhist}{Lookup table}

\item{labels}{vector of labels to include in plots. Options include: "Frequency", "TagNumber", and "VitFreq". If these labels are desired in plots, they MUST be included in your lookup table}

\item{fold}{Folder to save plots in}

\item{plotdataPath}{path to save plot data}
}
\value{
Returns plots for markdown in specified folder
}
\description{
Use movement data to create plots for Parturition Markdown Document
}
\examples{
\donttest{system.time({ vitMap(locdat=mdat3,vhist=vhist,labels = "Frequency", fold= plotfolder, plotdataPath = plotdatapath) })
)}

calculate movement rate
bring in frequency for report
rolling mean 
2-day quantile
GAUSSIAN BRIDGE METRICS
}
\keyword{VIT,}
\keyword{markdown}
\keyword{parturition,}
