% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheep.gps.R
\name{sheep.gps}
\alias{sheep.gps}
\title{sheep.gps}
\usage{
sheep.gps(keys, sheepdb, tzone, capcol, dateformat, mortcol, dnld.data,
  dnld.fold, extracols)
}
\arguments{
\item{keys}{full path to one or more key files}

\item{sheepdb}{path to sheep capture database}

\item{tzone}{desired time zone of gps data: "MST" or "US/Pacific"}

\item{capcol}{name of column where capture/start date exists}

\item{dateformat}{character string of the format that date columns are in}

\item{mortcol}{name of column where mortality date or end date exists}

\item{dnld.data}{Logical. TRUE/FALSE whether to include downloaded data from collars}

\item{dnld.fold}{path to root folder where downloaded data exists}

\item{extracols}{vector of extra columns that should be appended to GPS data. Names in vector MUST match names in lookup table}
}
\value{
Returns a data.frame with all gps data and any extra columns desired
}
\description{
Scrape GPS data  from online server, and merge with downloaded data from collars
}
\examples{

}
\keyword{ID,}
\keyword{animal}
\keyword{append}
\keyword{capture,}
\keyword{gps,}
