% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheep.gps.R
\name{sheep.gps}
\alias{sheep.gps}
\title{sheep.gps}
\usage{
sheep.gps(vecpath, sheepdb, tzone)
}
\arguments{
\item{vecpath}{path where vec data is located}

\item{sheepdb}{path to sheep capture database}
}
\value{
Returns a data.frame with all gps data, AnimalID, Sex, and Species
}
\description{
Merge Raw GPS Data with AnimalID, Sex, and Eartag ID
}
\examples{

}
\keyword{ID,}
\keyword{animal}
\keyword{append}
\keyword{capture,}
\keyword{gps,}
